+++ package/utils/IP.d.ts
@@ -1,5 +1,0 @@
-export default class IP {
-    static isIPv6(ip: string): boolean;
-    static normalizeIP(ip: string): Array<number>;
-    static getRandomIPv6(ip: string): string;
-}