+++ src/types/ytdl.ts
@@ -1,5 +1,6 @@
 import { YTDL_ClientTypes } from '@/meta/Clients';
 import { YT_Itag, YT_MicroformatRenderer, YT_NextApiResponse, YT_PlayerApiResponse, YT_Quality, YT_QualityLabel, YT_StreamingAdaptiveFormat, YT_Thumbnail, YT_VideoDetails } from './youtube';
+import { YTDL_GetInfoOptions } from './Options';
 
 export type YTDL_Author = {
     id: string;
@@ -56,10 +57,12 @@ export type YTDL_VideoDetailsAdditions = {
     published?: number | null;
 };
 
-export type YTDL_VideoDetails = Omit<YT_VideoDetails, 'lengthSeconds' | 'author' | 'thumbnail' | 'shortDescription' | 'viewCount'> & Omit<YT_MicroformatRenderer, 'title' | 'description'> & YTDL_VideoDetailsAdditions & {
-    lengthSeconds: number;
-    viewCount: number;
-};
+export type YTDL_VideoDetails = Omit<YT_VideoDetails, 'lengthSeconds' | 'author' | 'thumbnail' | 'shortDescription' | 'viewCount'> &
+    Omit<YT_MicroformatRenderer, 'title' | 'description'> &
+    YTDL_VideoDetailsAdditions & {
+        lengthSeconds: number;
+        viewCount: number;
+    };
 
 export type YTDL_RelatedVideo = {
     id: string;
@@ -70,7 +73,7 @@ export type YTDL_RelatedVideo = {
     isLive: boolean;
     published?: string;
     shortViewCountText?: string;
-    viewCount?: string;
+    viewCount?: number;
     lengthSeconds?: number;
 };
 
@@ -105,9 +108,6 @@ export type YTDL_VideoInfo = {
     relatedVideos: Array<YTDL_RelatedVideo>;
     formats: Array<YTDL_VideoFormat>;
     full: boolean;
-    isMinimumMode: boolean;
-    clients: Array<YTDL_ClientTypes>;
-    html5Player: string;
     live_chunk_readahead?: number;
     _playerApiResponses: {
         webCreator: YT_PlayerApiResponse | null;
@@ -120,6 +120,13 @@ export type YTDL_VideoInfo = {
     };
     _nextApiResponses: {
         web: YT_NextApiResponse | null;
+    };
+    _metadata: {
+        isMinimumMode: boolean;
+        clients: Array<YTDL_ClientTypes>;
+        html5Player: string;
+        id: string;
+        options: YTDL_GetInfoOptions;
     };
     _ytdl: {
         version: string;