+++ src/utils/Log.ts
@@ -1,25 +1,72 @@
 import { VERSION } from './constants';
 
+const OUTPUT_CONTROL_CHARACTER = {
+    red: '\x1b[31m',
+    green: '\x1b[32m',
+    yellow: '\x1b[33m',
+    blue: '\x1b[34m',
+    magenta: '\x1b[35m',
+
+    reset: '\x1b[0m',
+};
+
 export class Logger {
+    private static replaceColorTags(message: string): string {
+        message = message.replace(/<magenta>/g, OUTPUT_CONTROL_CHARACTER.magenta);
+        message = message.replace(/<\/magenta>/g, OUTPUT_CONTROL_CHARACTER.reset);
+        message = message.replace(/<debug>/g, OUTPUT_CONTROL_CHARACTER.magenta);
+        message = message.replace(/<\/debug>/g, OUTPUT_CONTROL_CHARACTER.reset);
+
+        message = message.replace(/<blue>/g, OUTPUT_CONTROL_CHARACTER.blue);
+        message = message.replace(/<\/blue>/g, OUTPUT_CONTROL_CHARACTER.reset);
+        message = message.replace(/<info>/g, OUTPUT_CONTROL_CHARACTER.blue);
+        message = message.replace(/<\/info>/g, OUTPUT_CONTROL_CHARACTER.reset);
+
+        message = message.replace(/<green>/g, OUTPUT_CONTROL_CHARACTER.green);
+        message = message.replace(/<\/green>/g, OUTPUT_CONTROL_CHARACTER.reset);
+        message = message.replace(/<success>/g, OUTPUT_CONTROL_CHARACTER.green);
+        message = message.replace(/<\/success>/g, OUTPUT_CONTROL_CHARACTER.reset);
+
+        message = message.replace(/<yellow>/g, OUTPUT_CONTROL_CHARACTER.yellow);
+        message = message.replace(/<\/yellow>/g, OUTPUT_CONTROL_CHARACTER.reset);
+        message = message.replace(/<warning>/g, OUTPUT_CONTROL_CHARACTER.yellow);
+        message = message.replace(/<\/warning>/g, OUTPUT_CONTROL_CHARACTER.reset);
+
+        message = message.replace(/<red>/g, OUTPUT_CONTROL_CHARACTER.red);
+        message = message.replace(/<\/red>/g, OUTPUT_CONTROL_CHARACTER.reset);
+        message = message.replace(/<error>/g, OUTPUT_CONTROL_CHARACTER.red);
+        message = message.replace(/<\/error>/g, OUTPUT_CONTROL_CHARACTER.reset);
+
+        return message;
+    }
+
+    private static convertMessage(message: string): string {
+        return this.replaceColorTags(message);
+    }
+
+    private static convertMessages(messages: Array<any>): Array<any> {
+        return messages.map((m) => this.replaceColorTags(m));
+    }
+
     public static debug(...messages: Array<any>) {
         if (VERSION.includes('dev') || VERSION.includes('beta') || VERSION.includes('test') || process.env.YTDL_DEBUG) {
-            console.log('\x1b[35m[  DEBUG  ]:\x1B[0m', ...messages);
+            console.log(this.convertMessage('<debug>[  DEBUG  ]:</debug>'), ...this.convertMessages(messages));
         }
     }
 
     public static info(...messages: Array<any>) {
-        console.info('\x1b[34m[  INFO!  ]:\x1B[0m', ...messages);
+        console.info(this.convertMessage('<info>[  INFO!  ]:</info>'), ...this.convertMessages(messages));
     }
 
     public static success(...messages: Array<any>) {
-        console.log('\x1b[32m[ SUCCESS ]:\x1B[0m', ...messages);
+        console.log(this.convertMessage('<success>[ SUCCESS ]:</success>'), ...this.convertMessages(messages));
     }
 
     public static warning(...messages: Array<any>) {
-        console.warn('\x1b[33m[ WARNING ]:\x1B[0m', ...messages);
+        console.warn(this.convertMessage('<warning>[ WARNING ]:</warning>'), ...this.convertMessages(messages));
     }
 
     public static error(...messages: Array<any>) {
-        console.error('\x1b[31m[  ERROR  ]:\x1B[0m', ...messages);
+        console.error(this.convertMessage('<error>[  ERROR  ]:</error>'), ...this.convertMessages(messages));
     }
 }